% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbspls_methods.R
\name{coef.mbspls}
\alias{coef.mbspls}
\title{Extract model coefficients from an mbspls object}
\usage{
\method{coef}{mbspls}(object, ...)
}
\arguments{
\item{object}{An mbspls object}

\item{...}{Further arguments (not used)}
}
\value{
A list with the weight vectors (coefficients) for each block
}
\description{
Extract model coefficients from an mbspls object
}
