% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{drop_singleton_dims}
\alias{drop_singleton_dims}
\title{Drop singleton dimensions from arrays}
\usage{
drop_singleton_dims(x)
}
\arguments{
\item{x}{An array or matrix}
}
\value{
The input with singleton dimensions removed
}
\description{
Helper function to remove singleton dimensions from arrays or matrices.
}
\keyword{internal}
