% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{mbs_data}
\alias{mbs_data}
\title{Create an MBSPLS data object from a list of matrices}
\usage{
mbs_data(..., metadata = list())
}
\arguments{
\item{...}{Named matrices to include in the MBSPLS analysis}

\item{metadata}{Optional list with additional metadata}
}
\value{
An object of class 'mbs_data'
}
\description{
Create an MBSPLS data object from a list of matrices
}
\examples{
# Create MBSPLS data from matrices
X1 <- matrix(rnorm(100*20), 100, 20)
X2 <- matrix(rnorm(100*30), 100, 30)
X3 <- matrix(rnorm(100*10), 100, 10)
data <- mbs_data(proteome = X1, metabolome = X2, clinical = X3)
}
