% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbspls_fit_fold.R
\name{mbspls_fit_fold}
\alias{mbspls_fit_fold}
\title{Fit MBSPLS on training data and project onto test data}
\usage{
mbspls_fit_fold(training_data, test_data, params, initial_weights = NULL)
}
\arguments{
\item{training_data}{List of training data matrices, each representing a block}

\item{test_data}{List of test data matrices, each representing a block}

\item{params}{List of MBSPLS parameters from mbspls_setup()}

\item{initial_weights}{Optional list of initial weight vectors}
}
\value{
A list containing:
\item{rho}{Correlation value or matrix norm between latent variables}
\item{weights}{Weight vectors for each block}
\item{latent_vars}{Latent variables for test data}
\item{training_latent_vars}{Latent variables for training data}
}
\description{
This function trains MBSPLS on training data and projects the test data
onto the resulting latent space.
}
\keyword{internal}
