% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbspls_methods.R
\name{predict.mbspls}
\alias{predict.mbspls}
\title{Predict method for mbspls objects}
\usage{
\method{predict}{mbspls}(object, newdata, ...)
}
\arguments{
\item{object}{An mbspls object}

\item{newdata}{A list of matrices (new data blocks)}

\item{...}{Further arguments (not used)}
}
\value{
A list containing projected latent variables and correlations
}
\description{
Project new data onto the MBSPLS latent space.
}
