% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{print.mbs_data}
\alias{print.mbs_data}
\title{Print method for mbs_data objects}
\usage{
\method{print}{mbs_data}(x, ...)
}
\arguments{
\item{x}{An mbs_data object}

\item{...}{Further arguments passed to or from other methods}
}
\value{
Invisibly returns the input object
}
\description{
Print method for mbs_data objects
}
