% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbspls_methods.R
\name{print.mbspls}
\alias{print.mbspls}
\title{Print method for mbspls objects}
\usage{
\method{print}{mbspls}(x, ...)
}
\arguments{
\item{x}{An mbspls object}

\item{...}{Further arguments passed to or from other methods}
}
\value{
Invisibly returns the input object
}
\description{
Print method for mbspls objects
}
