% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbspls_setup.R
\name{print.mbspls_params}
\alias{print.mbspls_params}
\title{Print MBSPLS parameters}
\usage{
\method{print}{mbspls_params}(x, ...)
}
\arguments{
\item{x}{An mbspls_params object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Invisibly returns the input object.
}
\description{
Print MBSPLS parameters
}
