% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbspls_methods.R
\name{print.summary.mbspls}
\alias{print.summary.mbspls}
\title{Print method for summary.mbspls objects}
\usage{
\method{print}{summary.mbspls}(x, ...)
}
\arguments{
\item{x}{A summary.mbspls object}

\item{...}{Further arguments passed to or from other methods}
}
\value{
Invisibly returns the input object
}
\description{
Print method for summary.mbspls objects
}
