% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{[.mbs_data}
\alias{[.mbs_data}
\title{Subset method for mbs_data objects}
\usage{
\method{[}{mbs_data}(x, i, j, ...)
}
\arguments{
\item{x}{An mbs_data object}

\item{i}{Indices for subsetting rows (samples)}

\item{j}{Block names or indices for subsetting columns (blocks)}

\item{...}{Not used}
}
\value{
A subset mbs_data object
}
\description{
Subset method for mbs_data objects
}
