% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbspls_methods.R
\name{summary.mbspls}
\alias{summary.mbspls}
\title{Summary method for mbspls objects}
\usage{
\method{summary}{mbspls}(object, ...)
}
\arguments{
\item{object}{An mbspls object}

\item{...}{Further arguments passed to or from other methods}
}
\value{
A list with summary statistics
}
\description{
Summary method for mbspls objects
}
